api/requirement.txt
streamlit
ollama
python-dotenv==1.0.1
gliner==0.2.13
datasets